+++ package.json
@@ -11,7 +11,9 @@
     "apollo-client": "^2.2.8",
     "apollo-link-batch-http": "^1.2.1",
     "apollo-link-error": "^1.0.7",
+    "b64-to-blob": "^1.2.19",
     "debounce-promise": "^3.1.0",
+    "file-saver": "^1.3.3",
     "formik": "^0.11.11",
     "graphql": "0.13.2",
     "graphql-tag": "^2.8.0",
